Refactored task list to use Svelte's universal reactivity
