+++ src/lib/List.svelte
@@ -1,26 +1,7 @@
 <script lang="ts">
+  import { tasks } from "./task-list.svelte";
+
   import Task from "./Task.svelte";
-
-  class Task_class {
-    public done: boolean = false;
-    public title: string = "";
-    public desc: string = "";
-
-    constructor(title: string, desc: string) {
-      this.title = title;
-      this.desc = desc;
-    }
-  }
-
-  let tasks = [
-    new Task_class("Task 1", "desc 1"),
-    new Task_class("Task 2", "desc 2"),
-    new Task_class("Task 3", "desc 3"),
-    new Task_class("Task 4", "desc 4"),
-  ];
-
-  // Todo: create list type
-  // Todo: create task type with title, description, done
 </script>
 
 <div class="list">
